[@wordpress/core-data] Move Schema from api-fetch
