+++ types/wordpress__core-data/actions.d.ts
@@ -1,4 +1,4 @@
-import { Schema } from '@wordpress/api-fetch';
+import { Schema } from './schema';
 import { Autosave, Entity } from './';
 
 /**
