+++ types/wordpress__editor/components/post-taxonomies/index.d.ts
@@ -1,4 +1,4 @@
-import { Schema } from '@wordpress/api-fetch';
+import { Schema } from '@wordpress/core-data';
 import { ComponentType, ReactNode } from 'react';
 
 declare namespace PostTaxonomies {
