+++ types/wordpress__editor/store/selectors.d.ts
@@ -1,4 +1,4 @@
-import { Schema } from '@wordpress/api-fetch';
+import { Schema } from '@wordpress/core-data';
 import { EditorSettings } from '@wordpress/block-editor';
 import { BlockInstance } from '@wordpress/blocks';
 import { Autosave } from '@wordpress/core-data';
